% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.factor.to.character.R
\name{convert.factor.to.character}
\alias{convert.factor.to.character}
\title{Convert factor to characters}
\usage{
convert.factor.to.character(x, ...)
}
\arguments{
\item{x:}{a vector of any kind.  If x is a factor, it will be converted to a character vector.}
}
\description{
function that converts factors to character
}
