% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.filter.R
\name{dt.filter}
\alias{dt.filter}
\title{dat:  a data.frame
the.filter:  a character value or expression stating the logical operations to be performed.
A subset of the data corresponding to the rows on which the filtering statement are TRUE is returned.}
\usage{
dt.filter(dat, the.filter = NULL)
}
\description{
dat:  a data.frame
the.filter:  a character value or expression stating the logical operations to be performed.
A subset of the data corresponding to the rows on which the filtering statement are TRUE is returned.
}
