% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.first.k.rows_and_dt.last.k.rows.R
\name{dt.first.k.rows}
\alias{dt.first.k.rows}
\title{dt.first.k.rows}
\source{
DTwrappers::dt.choose.cols
}
\usage{
dt.first.k.rows(
  dt.name,
  k = NULL,
  the.variables = ".",
  the.filter = NULL,
  grouping.variables = NULL,
  sortby.group = TRUE,
  return.as = "result",
  envir = .GlobalEnv,
  ...
)
}
\arguments{
\item{the.filter}{a character value, logical vector, or expression stating the logical operations used to filter the data.  See create.filter.expression for details.  The filtering step will be applied prior to generating the counts.  Defaults to NULL unless otherwise specified.}

\item{grouping.variables}{a character vector specifying the variables to group by in the calculation.  Only variables in the data will be used.  When NULL, no grouping will be performed.}

\item{return.as}{a character value specifying what output should be returned.  return.as = "result" provides the resulting table.  return.as = "code" provides a data.table coding statement that can generate the resulting table.  return.as = "all" provides both the resulting table and the code.  If the coding statement was specified using calls to get() or eval(), then both an original.statement and the resulting code (a translated statement from the getDTeval package) will be provided.}

\item{envir}{the environment in which the code would be evaluated; .GlobalEnv by default.}

\item{...}{additional arguments to be passed}

\item{grouping.type}{a character value specifying whether the grouping should be sorted ("keyby") or as is ("by").  Defaults to "keyby" unless "by" is specified.}

\item{count.name}{a character value specifying the name of the column of counts in the resulting table.  This value defaults to "N" unless otherwise specified.}
}
\description{
This function returns the first k rows from the given data.  It is built as a wrapper function of data.table's filter (the i step).  This calculation can be specified either overall or in groups.  A filter can also be applied so that only qualifying values would be considered.  A subset of the variables may also be selected.  Options also exist to return a data.table coding statement (result = "code") for educational purposes or both the result and the code together (result = "all").  For examples, please see the vignette.

#' @param dt.name a character value specifying the name of a data.frame or data.table object to select data from.
}
\note{
Calls dt.choose.cols.R with first.k = k.
}
\examples{

}
