% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.remove.variables.R
\name{dt.remove.variables}
\alias{dt.remove.variables}
\title{dt.remove.variables}
\usage{
dt.remove.variables(dat, the.variables)
}
\arguments{
\item{dat}{a data.frame object}

\item{the.variables}{a character vector with the column names to be removed from dat.}
}
\description{
TBD
}
