% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.summarize.R
\name{dt.summarize}
\alias{dt.summarize}
\title{dt Summary}
\usage{
dt.summarize(dat, the.variables = ".", the.functions = c("min",
  "lower.quartile", "median", "mean", "upper.quartile", "max", "sd",
  "num.records", "total.missing"), na.rm = TRUE, the.filter = NULL,
  grouping.variables = NULL, grouping.type = "keyby",
  add.function.name = TRUE, include.total = TRUE, format.as = "wide",
  ...)
}
\arguments{
\item{dat}{a data.frame object.}

\item{the.variables}{a character vector specifying the variables that we want to apply a function to.
Only values that exist in names(dat) will be used; other values in the.variables will be excluded from the calculation.  
When the.variables includes ".", then all values in names(dat) will be selected.  
Values of the.variables that also exist in grouping.variables will be excluded from the.variables (but grouped by these values).}

\item{the.functions}{a character vector or list specifying the name of the function to apply to the variables.  This may either be specified by the name of the function as a character (e.g. "mean") or by defining a function; e.g. function(x){return(mean(x = x))} that can be computed on each column of data as specified in the.variables.}

\item{na.rm}{a logical value specifying whether missing values should be removed from the calculations specified by the.functions.}

\item{the.filter}{a character value, logical value, or expression stating the logical operations to be performed in filtering the data prior to calculating the.function.}

\item{grouping.variables}{a character vector specifying variables to group by in performing the computation.  Only values that exist in names(dat) will be used.}

\item{grouping.type}{a character value specifying whether the grouping should be sorted (keyby) or as is (by).  Defaults to keyby unless "by" is specified.}

\item{add.function.name}{a logical value specifying whether the name of the function applied should be appended to the column names in the resulting table.}

\item{include.total}{a logical value specifying whether the.functions should be applied both to the entire table and to the groups specified in grouping.variables.}

\item{format.as}{a character value with "long" corresponding to long format.  All other values of format.as will produce results in wide format.}
}
\description{
dt Summary
}
