% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTwrapers-package.R
\docType{package}
\name{DTwrapers-package}
\alias{DTwrapers}
\alias{DTwrapers-package}
\title{DTwrapers: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Anderson Nelson \email{an2908@columbia.edu} [contributor]

Authors:
\itemize{
  \item David Shilane \email{david.shilane@columbia.edu}
}

Other contributors:
\itemize{
  \item Caffrey Lee \email{cl3802@columbia.edu} [contributor]
  \item Zichen Huang \email{zh2380@columbia.edu} [contributor]
}

}
\keyword{internal}
