% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTwrapers-package.R
\docType{package}
\name{DTwrappers-package}
\alias{DTwrappers}
\alias{DTwrappers-package}
\title{DTwrappers: Wrapper for the data.table package}
\description{
Many newcomers to the data.table packages are overwhelmed by the unique syntax and DTwrappers contains wrapper functions for data.table while still maintaining the data.table execution speed.
}
\author{
\strong{Maintainer}: Caffrey Lee \email{cl3802@columbia.edu} [contributor]

Authors:
\itemize{
  \item David Shilane \email{david.shilane@columbia.edu}
}

Other contributors:
\itemize{
  \item Anderson Nelson \email{an2908@columbia.edu} [contributor]
  \item Zichen Huang \email{zh2380@columbia.edu} [contributor]
}

}
\keyword{internal}
