% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.vector.expression.R
\name{reduce.vector.expression}
\alias{reduce.vector.expression}
\title{Takes a numeric vector and produces a statement with a more compact representation.  For instance, c(1,2,3,4) would become '1:4' and c(1:3, 4:6) could become '1:6'.}
\usage{
reduce.vector.expression(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
None
}
\description{
Takes a numeric vector and produces a statement with a more compact representation.  For instance, c(1,2,3,4) would become '1:4' and c(1:3, 4:6) could become '1:6'.
}
