% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.numerics.R
\name{var.numerics}
\alias{var.numerics}
\title{var numerics
x:  a vector. 
na.rm:  a logical value specifying whether missing values should be removed from the calculations specified by the.functions.}
\usage{
var.numerics(x, na.rm = TRUE, non.numeric.value = "missing", ...)
}
\description{
non.numeric.value:if "missing", returns NA for variables that are not numeric, integer, logical, or complex.  Otherwise returns first entry of the vector.
If x is numeric, integer, logical, or complex, then its variance will be computed.  Otherwise, the first value of x will be returned untouched.
}
