% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.filter.expression.R
\name{create.filter.expression}
\alias{create.filter.expression}
\title{Internal function}
\usage{
create.filter.expression(the.filter)
}
\description{
the.filter:  a character value, logical value, or expression stating the logical operations to be performed in filtering the data prior to calculating the.function.
}
