% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.filter.expression.R
\name{create.filter.expression}
\alias{create.filter.expression}
\title{create.filter.expression}
\usage{
create.filter.expression(the.filter)
}
\arguments{
\item{the.filter}{The filter criteria could be a character value, logical value, or expression stating the logical operations to be performed in filtering the data.}
}
\description{
A function that creates filter expressions. Expression could be a character value, logical value, or expression stating the logical operations.
}
\examples{
 
age.name = "Age"
region.name = "Region"
 
create.filter.expression(the.filter = NULL)
create.filter.expression(the.filter = c(age.name, region.name))
create.filter.expression(the.filter = age.name == region.name)
create.filter.expression(the.filter = "get(region.name) == 'South'")
create.filter.expression(the.filter = "get(age.name) > 20")
create.filter.expression(the.filter = "get(region.name) == 'South' & get(age.name) > 20")
 
}
