% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.convert.factor.to.character.R
\name{dt.convert.factor.to.character}
\alias{dt.convert.factor.to.character}
\title{dat:  a data.frame object.
the.variables:  a character vector specifying the variables that we want to apply a function to.  Only values that exist in names(dat) will be used; other values in the.variables will be excluded from the calculation.  When the.variables includes ".", then all values in names(dat) will be selected.
add.function.name:  a logical value.  If TRUE, then the names of the variables are concatenated with the suffix convert.factor.to.character.}
\usage{
dt.convert.factor.to.character(dat, the.variables = ".",
  add.function.name = FALSE, ...)
}
\description{
dat:  a data.frame object.
the.variables:  a character vector specifying the variables that we want to apply a function to.  Only values that exist in names(dat) will be used; other values in the.variables will be excluded from the calculation.  When the.variables includes ".", then all values in names(dat) will be selected.
add.function.name:  a logical value.  If TRUE, then the names of the variables are concatenated with the suffix convert.factor.to.character.
}
