% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.define.variable.R
\name{dt.define.variable}
\alias{dt.define.variable}
\title{dat:  a data.frame object
variable.name:  a character value specifying the name of the new column
the.values:  a vector of data specifying the values of the new column.
Note:  the data.frame dat will be converted to a data.table object to facilitate adding the new column by reference (e.g. efficiently with regard to the usage of memory)}
\usage{
dt.define.variable(dat, variable.name, the.values,
  specification = "by.value", the.filter = NULL)
}
\description{
dat:  a data.frame object
variable.name:  a character value specifying the name of the new column
the.values:  a vector of data specifying the values of the new column.
Note:  the data.frame dat will be converted to a data.table object to facilitate adding the new column by reference (e.g. efficiently with regard to the usage of memory)
}
